AXIS_AI/requirements.txt
speechrecognition
pyttsx3
pyaudio
openai
python-dotenv
pyautogui
psutil
